20220813-1750
Fuelfix, merge

+++ intsstrings.c
@@ -20,15 +20,15 @@ int OVERTERRAIN_TYPE = 5;
 int OVERTERRAIN_SUBTYPE = 4;
 int CheckTime = 1;
 
-int FuelCost = 0;
-int FSOneXMin = 10;
-int FSOneXMax = 0;
-int FSOneZMin = 10;
-int FSOneZMax = 0;
-int FSTwoXMin = 10;
-int FSTwoXMax = 0;
-int FSTwoZMin = 10;
-int FSTwoZMax = 0;
+float FuelCost = 0;
+float FSOneXMin = 10;
+float FSOneXMax = 0;
+float FSOneZMin = 10;
+float FSOneZMax = 0;
+float FSTwoXMin = 10;
+float FSTwoXMax = 0;
+float FSTwoZMin = 10;
+float FSTwoZMax = 0;
 
 /* TERRAINS */
 const int SandC = 1;
